file upload and download section completed

+++ NodeJS The Complete Guide/controllers/admin.js
@@ -1,5 +1,9 @@
 const { validationResult } = require('express-validator/check');
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
+const fs = require('fs');
+const path = require('path');
+const rootDir = require('../util/path');
 
 const Product = require('../models/product');
 
@@ -18,8 +22,27 @@ exports.getAddProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = (req, res, next) => {
-    const { title, imageUrl, description, price } = req.body;
+    const title = req.body.title;
+    const description = req.body.description;
+    const price = req.body.price;
+    const image = req.file;
     const errors = validationResult(req);
+
+    if(!image) {
+       return res.status(422).render('admin/edit-product', {
+            pageTitle: 'Add Product',
+            path: '/admin/edit-product',
+            editing: false,
+            hasError: true,
+            product: {
+                title: title,
+                price: price,
+                description: description
+            },
+            errorMessage: "Attached file is not an image!",
+            validationErrors: []
+        });
+    }
 
     if (!errors.isEmpty()) {
         console.log(errors.array());
@@ -30,7 +53,6 @@ exports.postAddProduct = (req, res, next) => {
             hasError: true,
             product: {
                 title: title,
-                imageUrl: imageUrl,
                 price: price,
                 description: description
             },
@@ -40,11 +62,11 @@ exports.postAddProduct = (req, res, next) => {
     }
 
     const product = new Product({
-        _id: new mongoose.Types.ObjectId('68452b51ffc21ec548102e88'),
+        //_id: new mongoose.Types.ObjectId('68452b51ffc21ec548102e88'),
         title : title,
         price: price,
         description : description,
-        imageUrl : imageUrl,
+        imageUrl : 'images/' + image.filename,
         userId : req.user
     });
     product
@@ -105,24 +127,31 @@ exports.getEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id : prodId, userId: req.user._id})
-    .then(() => {
-        console.log('DESTROYED PRODUCT');
-        res.redirect('/admin/products');
-    })
-    .catch(err => {
-        const error = new Error(err);
-        error.httpStatusCode = 500;
-        return next(error);  
-    });
-    
+    Product.findById(prodId)
+        .then(product => {
+            if(!product) {
+                return next(new Error('Product not found.'));
+            }
+            const imagePath = path.join(rootDir, '..', product.imageUrl);
+            fileHelper.deleteFile(imagePath);
+            return Product.deleteOne({ _id : prodId, userId: req.user._id});
+        })
+        .then(() => {
+            console.log('DESTROYED PRODUCT');
+            res.redirect('/admin/products');
+        })
+        .catch(err => {
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);  
+        });
 };
 
 exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDesc = req.body.description;
 
     const errors = validationResult(req);
@@ -135,7 +164,6 @@ exports.postEditProduct = (req, res, next) => {
           hasError: true,
           product: {
             title: updatedTitle,
-            imageUrl: updatedImageUrl,
             price: updatedPrice,
             description: updatedDesc,
             _id: prodId
@@ -153,7 +181,11 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        product.imageUrl = updatedImageUrl;
+        if(image) {
+            const imagePath = path.join(rootDir, '..', product.imageUrl);
+            fileHelper.deleteFile(imagePath);
+            product.imageUrl = 'images/' + image.filename
+        }
         return product.save()
         .then(result => {
             console.log("UPDATED PRODUCT!");
